Move gameday controllers to its own py file
